(fix) Now if clear all local storage - example cards doesnt dissappear
