+++ src/App.tsx
@@ -4,12 +4,6 @@ import EnterNamePopup from "./components/Popup";
 import "bootstrap/dist/css/bootstrap.min.css";
 import store from "./utils/store";
 import Board from "./components/Board";
-import {
-  DONE_COLUMN_NAME,
-  IN_PROGRESS_COLUMN_NAME,
-  TESTING_COLUMN_NAME,
-  TODO_COLUMN_NAME,
-} from "./constants/exampleColumnNames";
 import { ColumnsProvider } from "./context/ColumnsContext";
 
 function App() {
@@ -20,17 +14,6 @@ function App() {
 
     if (!author) {
       setIsShowPopup(true);
-    }
-
-    const columns = store.getColumns();
-
-    if (!columns.length) {
-      store.addColumns([
-        TODO_COLUMN_NAME,
-        IN_PROGRESS_COLUMN_NAME,
-        TESTING_COLUMN_NAME,
-        DONE_COLUMN_NAME,
-      ]);
     }
   }, []);
