+++ src/context/ColumnsContext.tsx
@@ -1,4 +1,10 @@
 import React, { useEffect, useMemo, useState } from "react";
+import {
+  DONE_COLUMN_NAME,
+  IN_PROGRESS_COLUMN_NAME,
+  TESTING_COLUMN_NAME,
+  TODO_COLUMN_NAME,
+} from "../constants/exampleColumnNames";
 import { IColumn } from "../entities/Column/IColumn";
 import store from "../utils/store";
 
@@ -18,6 +24,17 @@ export const ColumnsProvider = (props: any) => {
   const [columns, setColumns] = useState<IColumn[]>([]);
 
   useEffect(() => {
+    const columns = store.getColumns();
+
+    if (columns.length === 0) {
+      store.addColumns([
+        TODO_COLUMN_NAME,
+        IN_PROGRESS_COLUMN_NAME,
+        TESTING_COLUMN_NAME,
+        DONE_COLUMN_NAME,
+      ]);
+    }
+
     setColumns(store.getColumns());
   }, []);
